// src/interfaces/characterInterface.tsx
// Generated by https://quicktype.io

export interface CharacterInterfaceDB {
  info: Info[];
  results: Pers[];
}

export interface Info {
  count: number;
  pages: number;
  next: string;
  prev: null;
}

export interface Pers {
  id: number;
  name: string;
  status: Status;
  species: Species;
  type: string;
  gender: Gender;
  origin: Location;
  location: Location;
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export enum Gender {
  Female = 'Female',
  Male = 'Male',
  Unknown = 'unknown',
}

export interface Location {
  name: string;
  url: string;
}

export enum Species {
  Alien = 'Alien',
  Human = 'Human',
}

export enum Status {
  Alive = 'Alive',
  Dead = 'Dead',
  Unknown = 'unknown',
}